### 2023-08-18 #5768 @editorialbot
[PRE REVIEW]: greeks: Sensitivities of Prices of Financial Options and Implied Volatilities
**Submitting author:** @ahudde (<a href="http://orcid.org/0000-0002-5652-2815">Anselm Hudde</a>)
**Repository:** https://github.com/ahudde/greeks
**Branch with paper.md** (empty if default branch):
**Version:** 1.2.0
