### 2023-08-28 28a9b14
Update README.md
Added R-CMD-CHECK-badge in README.md
