+++ README.md
@@ -1,4 +1,5 @@
 # greeks
+[![R-CMD-check](https://github.com/ahudde/greeks/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/ahudde/greeks/actions/workflows/R-CMD-check.yaml)
 ![](https://cranlogs.r-pkg.org/badges/greeks?color=brightgreen)
 ![](https://cranlogs.r-pkg.org/badges/grand-total/greeks?color=brightgreen)
 ![](https://www.r-pkg.org/badges/version-ago/greeks)
