+++ tests/testthat/test-Implied_Volatility.R
@@ -1,0 +1,72 @@
+test_that("implied volatility is correct", {
+  # The implied volatility is computed with different parameters, one of them
+  # being `option_price`. The resulting `implied_volatility` is the used to
+  # compute the `option_price_tets`. Both values should be very close
+
+  number_of_runs <- 2
+
+  set.seed(42)
+  initial_price <- runif(number_of_runs, 90, 110)
+  exercise_price <- runif(number_of_runs, 90, 110)
+  r <- runif(number_of_runs, -0.01, 0.1)
+  time_to_maturity <- runif(number_of_runs, 0.2, 6)
+  dividend_yield <- runif(number_of_runs, 0, 0.1)
+  model <- "Black_Scholes"
+  option_type <- c(rep(c("European"), 2))
+  payoff <- rep(c("call", "put"), 1)
+  volatility_to_compute_initial_price <- runif(number_of_runs, 0.01, 1)
+
+  option_price <- numeric(number_of_runs)
+  option_price_test <- numeric(number_of_runs)
+
+  for(i in 1:number_of_runs) {
+    option_price[i] <-
+      Greeks(
+        initial_price = initial_price[i],
+        exercise_price = exercise_price[i],
+        r = r[i],
+        time_to_maturity = time_to_maturity[i],
+        volatility = volatility_to_compute_initial_price[i],
+        dividend_yield = dividend_yield[i],
+        option_type = option_type[i],
+        payoff = payoff[i],
+        greek = "fair_value"
+      )
+  }
+
+  implied_volatility <- numeric(number_of_runs)
+
+  for(i in 1:number_of_runs) {
+
+    implied_volatility[i] <-
+      Implied_Volatility(
+        option_price = option_price[i],
+        initial_price = initial_price[i],
+        exercise_price = exercise_price[i],
+        r = r[i],
+        time_to_maturity = time_to_maturity[i],
+        dividend_yield = dividend_yield[i],
+        model = model,
+        option_type = option_type[i],
+        payoff = payoff[i]
+      )
+
+    option_price_test[i] <-
+      Greeks(
+        initial_price = initial_price[i],
+        exercise_price = exercise_price[i],
+        r = r[i],
+        time_to_maturity = time_to_maturity[i],
+        volatility = implied_volatility[i],
+        dividend_yield = dividend_yield[i],
+        model = model,
+        option_type = option_type[i],
+        payoff = payoff[i],
+        greek = "fair_value"
+      )
+
+  }
+
+  expect(max(abs(option_price_test - option_price)) < 1e-06)
+
+})
