+++ tests/testthat/test-Implied_Volatility.R
@@ -1,10 +1,12 @@
 test_that("implied volatility is correct", {
+
   # The implied volatility is computed with different parameters, one of them
   # being `option_price`. The resulting `implied_volatility` is the used to
-  # compute the `option_price_tets`. Both values should be very close
+  # compute the `option_price_test`. Both values should be very close.
 
   number_of_runs <- 2
 
+  # the parameters are chosen at random
   set.seed(42)
   initial_price <- runif(number_of_runs, 90, 110)
   exercise_price <- runif(number_of_runs, 90, 110)
@@ -12,12 +14,15 @@ test_that("implied volatility is correct", {
   time_to_maturity <- runif(number_of_runs, 0.2, 6)
   dividend_yield <- runif(number_of_runs, 0, 0.1)
   model <- "Black_Scholes"
-  option_type <- c(rep(c("European"), 2))
-  payoff <- rep(c("call", "put"), 1)
+  option_type <- c("European", "European")
+  payoff <- c("call", "put")
   volatility_to_compute_initial_price <- runif(number_of_runs, 0.01, 1)
 
   option_price <- numeric(number_of_runs)
+  option_price_zero_vol <- numeric(number_of_runs)
   option_price_test <- numeric(number_of_runs)
+
+  implied_volatility <- numeric(number_of_runs)
 
   for(i in 1:number_of_runs) {
     option_price[i] <-
@@ -32,11 +37,25 @@ test_that("implied volatility is correct", {
         payoff = payoff[i],
         greek = "fair_value"
       )
-  }
 
-  implied_volatility <- numeric(number_of_runs)
+    option_price_zero_vol[i] <-
+      Greeks(
+        initial_price = initial_price[i],
+        exercise_price = exercise_price[i],
+        r = r[i],
+        time_to_maturity = time_to_maturity[i],
+        volatility = 0,
+        dividend_yield = dividend_yield[i],
+        payoff = payoff[i],
+        greek = "fair_value"
+      )
 
-  for(i in 1:number_of_runs) {
+    # We check if the implied volatility for the chosen parameters is
+    # well-defined
+    if (option_price[i] <= option_price_zero_vol[i] + 1e-12) {
+      option_price[i] <- 0
+      next
+    }
 
     implied_volatility[i] <-
       Implied_Volatility(
