### 2023-08-31 c22c893
selection of the faster function BS_Malliavin_Asian_Greeks in more cases
