### 2023-08-31 bb0eec7
Volatility for option_price_zero_vol changed from 0 to 1e-12 to avoid numerical problems
