### 2023-09-03 5e05985
Made code more DRY
