+++ R/BS_European_Greeks.R
@@ -57,9 +57,9 @@ BS_European_Greeks <-
 
     dnorm_d1 <- dnorm(d1)
 
-    if (payoff == "call") {
+    pnorm_d2 <- pnorm(d2)
 
-      pnorm_d2 <- pnorm(d2)
+    if (payoff == "call") {
 
       ## option-value
 
@@ -75,11 +75,6 @@ BS_European_Greeks <-
       if ("delta" %in% greek) {
         result['delta'] <- exp_minus_dividend_yield_times_time_to_maturity *
           pnorm_d1
-      }
-      if ("vega" %in% greek) {
-        result['vega'] <-
-          initial_price * exp_minus_dividend_yield_times_time_to_maturity *
-          dnorm_d1 * sqrt_time_to_maturity
       }
       if ("theta" %in% greek) {
         result['theta'] <-
@@ -98,6 +93,87 @@ BS_European_Greeks <-
         result['epsilon'] <-
           -initial_price * time_to_maturity *
           exp_minus_dividend_yield_times_time_to_maturity * pnorm_d1
+      }
+
+      ## second-order Greeks
+
+      if ("charm" %in% greek) {
+        result["charm"] <-
+          (dividend_yield * exp_minus_dividend_yield_times_time_to_maturity *
+             pnorm_d1) -
+          (exp_minus_dividend_yield_times_time_to_maturity * dnorm_d1 *
+             (2 * (r - dividend_yield) * time_to_maturity -
+                d2 * volatility * sqrt_time_to_maturity) /
+             (2 * time_to_maturity * volatility * sqrt_time_to_maturity))
+      }
+    }
+
+    if (payoff == "put") {
+
+      pnorm_minus_d2 <- pnorm(-d2)
+
+      dnorm_minus_d2 <- dnorm(-d2)
+
+      pnorm_minus_d1 <- pnorm(-d1)
+
+      ## option-value
+
+      if ("fair_value" %in% greek) {
+        result['fair_value'] <-
+          exp_minus_r_times_time_to_maturity * exercise_price * pnorm_minus_d2 -
+          initial_price * exp_minus_dividend_yield_times_time_to_maturity *
+          pnorm_minus_d1
+      }
+
+      ## first-order Greeks
+
+      if ("delta" %in% greek) {
+        result['delta'] <-
+          -exp_minus_dividend_yield_times_time_to_maturity * pnorm_minus_d1
+      }
+      if ("theta" %in% greek) {
+        result['theta'] <-
+          -initial_price * dnorm_d1 * volatility *
+          exp_minus_dividend_yield_times_time_to_maturity /
+          (2 * sqrt_time_to_maturity) - dividend_yield * initial_price *
+          pnorm_minus_d1 * exp_minus_dividend_yield_times_time_to_maturity +
+          r * exercise_price * exp_minus_r_times_time_to_maturity *
+          pnorm_minus_d2
+      }
+      if ("rho" %in% greek) {
+        result['rho'] <-
+          -exercise_price * time_to_maturity *
+          exp_minus_r_times_time_to_maturity * pnorm_minus_d2
+      }
+      if ("epsilon" %in% greek) {
+        result['epsilon'] <-
+          initial_price * time_to_maturity *
+          exp_minus_dividend_yield_times_time_to_maturity * pnorm_minus_d1
+      }
+
+      # second-order Greeks
+
+      if ("charm" %in% greek) {
+        result["charm"] <-
+          (-dividend_yield * exp_minus_dividend_yield_times_time_to_maturity *
+             pnorm_minus_d1) -
+          (exp_minus_dividend_yield_times_time_to_maturity * dnorm_d1 *
+             (2 * (r - dividend_yield) * time_to_maturity -
[…]
