### 2023-09-03 13a3ff5
test-BS-European_Greekes tests veta and color
