### 2023-09-04 f2c1da7
make sure NA is returned for Greeks that not implemented
