+++ .github/workflows/codecov.yml
@@ -1,4 +1,13 @@
-- name: Upload coverage reports to Codecov
+on:
+  push:
+    branches: [main, master]
+  pull_request:
+    branches: [main, master]
+
+name: codecov
+
+jobs:
+  name: Upload coverage reports to Codecov
   uses: codecov/codecov-action@v3
   env:
     CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
