+++ .github/workflows/R-CMD-check.yaml
@@ -6,11 +6,6 @@ on:
   pull_request:
     branches: [main, master]
 
-name: Upload coverage reports to Codecov
-  uses: codecov/codecov-action@v3
-  env:
-    CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
-
 name: R-CMD-check
 
 jobs:
