+++ .github/workflows/workflow.yml
@@ -1,5 +1,5 @@
 steps:
-- uses: actions/checkout@grees_1.3.0
+- uses: actions/checkout@greeks_1.3.0
 - uses: codecov/codecov-action@v3
   with:
     token: ${{ secrets.CODECOV_TOKEN }}
