+++ .github/workflows/workflow.yml
@@ -2,13 +2,6 @@ name: Example workflow for Codecov
 on: [push]
 jobs:
   run:
-    runs-on: ${{ matrix.os }}
-    strategy:
-      matrix:
-        os: [ubuntu-latest, macos-latest, windows-latest]
-    env:
-      OS: ${{ matrix.os }}
-      PYTHON: '3.10'
     steps:
     - uses: actions/checkout@master
     - name: Setup Python
