+++ .github/workflows/workflow.yml
@@ -1,14 +1,0 @@
-name: Example workflow for Codecov
-on: [push]
-jobs:
-  R-CMD-check:
-    runs-on: ${{ matrix.config.os }}
-steps:
-- uses: actions/checkout@greeks_1.3.0
-- uses: codecov/codecov-action@v3
-  with:
-    files: ./coverage1.xml,./coverage2.xml # optional
-    flags: unittests # optional
-    name: codecov-umbrella # optional
-    fail_ci_if_error: true # optional (default = false)
-    verbose: true # optional (default = false)
