+++ .github/workflows/workflow.yml
@@ -1,3 +1,9 @@
+on:
+  push:
+    branches: [main, master]
+  pull_request:
+    branches: [main, master]
+
 steps:
 - uses: actions/checkout@main
 - uses: codecov/codecov-action@v3
