+++ .github/workflows/workflow.yml
@@ -4,6 +4,11 @@ on:
   pull_request:
     branches: [main, master]
 
+name: codecov
+
+jobs:
+  codecov:
+
 steps:
 - uses: actions/checkout@main
 - uses: codecov/codecov-action@v3
