+++ .github/workflows/workflow.yml
@@ -10,13 +10,6 @@ jobs:
   codecov:
     runs-on: ${{ matrix.config.os }}
 
-    steps:
-      - uses: actions/checkout@main
-      - uses: codecov/codecov-action@v3
-      with:
-        token: ${{ secrets.CODECOV_TOKEN }}
-        files: ./coverage1.xml,./coverage2.xml # optional
-        flags: unittests # optional
-        name: codecov-umbrella # optional
-        fail_ci_if_error: true # optional (default = false)
-        verbose: true # optional (default = false)
+steps:
+- uses: actions/checkout@main
+- uses: codecov/codecov-action@v1
