### 2023-09-05 e3b3038
Greeks.R: More DRY
