### 2023-09-06 f9c2183
Moved function to test-Binomial_American_Greeks.R
