### 2023-09-07 41c5a65
Removed numerical instability
