### 2023-09-14 c67cac2
Improved Binomial_American_Greek
