+++ src/Binomial_American_Greeks.cpp
@@ -65,7 +65,7 @@ NumericVector Binomial_American_Greeks_cpp(double initial_price = 100,
     american_value(i) = payoff_function(0.0, price(2*steps - 2*i), exercise_price);
   }
 
-  european_value = american_value;
+  european_value = clone(american_value);
 
   for(j = steps-1; j >= 0; j--) {
     for(i = 0; i <= j; i++) {
@@ -73,7 +73,7 @@ NumericVector Binomial_American_Greeks_cpp(double initial_price = 100,
         payoff_function(
           exp_min_r_dt * (p*american_value(i) + q*american_value(i+1)),
           price(2*steps - 2*i + j - steps), exercise_price);
-      european_value(i) = p * european_value(i) + q * european_value(i);
+      european_value(i) = p * european_value(i) + q * european_value(i+1);
     }
   }
 
