### 2023-10-16 #5768 @rmcd1024
@csoneson I've never reviewed for JOSS --- what turnaround are you looking
for? Is it just a matter of going through the checklist?

On Fri, Oct 13, 2023 at 9:23 AM Charlotte Soneson ***@***.***>
wrote:

> 👋🏻 @rmcd1024 <https://github.com/rmcd1024>, @ArdiaD
> <https://github.com/ArdiaD>, @syoh <https://github.com/syoh> - would you
> be interested in reviewing this submission for the Journal of Open Source
> Software (JOSS)?
>
> greeks: Sensitivities of Prices of Financial Options and Implied
> Volatilities
> (#5768 <https://github.com/openjournals/joss-reviews/issues/5768>)
>
> JOSS is a free, open-source, community-driven and developer-friendly
> online journal. The review process is unique: it takes place in a GitHub
> issue, is open, and author-reviewer-editor conversations are encouraged.
>
> JOSS reviews involve downloading and installing the software, and
> inspecting the repository and submitted paper for key elements. See
> https://joss.readthedocs.io/en/latest/review_criteria.html.
>
> Thanks in advance!
>
