### 2023-10-17 #5768 @csoneson
@rmcd1024 We usually ask reviewers to provide an initial assessment within 2-3 weeks. In many cases it may be easier to raise points and provide comments in the associated GitHub issue as one goes through the submission, rather than posting all comments in one go. The checklist contains the aspects that we ask reviewers to check for in the submission - some of the points can be quickly checked, but others of course involve installing and testing the software.
