### 2023-10-17 #5768 @syoh
@csoneson, thanks for the invitation. I have been wanting to do a review for JOSS. Unfortunately, my hands are full until next April, but I am happy to review something then.
