### 2023-10-25 #5768 @csoneson
👋🏻 @bahung, @edrictam - would one of you be interested in reviewing this submission for the Journal of Open Source Software (JOSS)?

> greeks: Sensitivities of Prices of Financial Options and Implied Volatilities
(https://github.com/openjournals/joss-reviews/issues/5768)

JOSS is a free, open-source, community-driven and developer-friendly online journal. The review process is unique: it takes place in a GitHub issue, is open, and author-reviewer-editor conversations are encouraged.

JOSS reviews involve downloading and installing the software, and inspecting the repository and submitted paper for key elements. See https://joss.readthedocs.io/en/latest/review_criteria.html.

Thanks in advance!
