### 2023-10-25 #5987 @editorialbot
[REVIEW]: greeks: Sensitivities of Prices of Financial Options and Implied Volatilities
**Submitting author:** @ahudde (<a href="http://orcid.org/0000-0002-5652-2815">Anselm Hudde</a>)
**Repository:** https://github.com/ahudde/greeks
**Branch with paper.md** (empty if default branch):
**Version:** greeks_1.4.0
