### 2023-10-30 #5987 @bahung
@ahudde could you please update the LICENSE that complies with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license?
