### 2023-10-30 #5987 @ahudde
Dear @bahung thank you very much, I have changed the license to MIT license.
