### 2023-10-30 anselmhudde/greeks#5 @lrnv
[docs] Lacking content, description of context and references
The documentations of the different functions are a bit lacking... In particular, the "Details" sections, the "Value" section, that are supposed to give an overview of what the function will do / does. Without asking too much of you, maybe you could include the mathematical definitions of the objects that your functions are computing and a few references to the corresponding litterature ?

As an exemple, let's take the `Greeks` function documentation:

````
Description : Computes the Greeks of various options
Value : Named vector containing the values of the Greeks specified in the parameter greek.
````

The description field could include a discussion of :
- What are the greeks ? Give a broad overview and link to the litterature for more details, this will be enough : " "Greeks are derivative of the value function of assets, usually used for blablabla...". You may even link wikipedia !
- Under which model / assumptions your code works ?
- All that with references to the litterature. IIRC there is the possiiblity to cite the litterature in R docs.

Same remark for all functions... One possibility is to have this done in a "master function" and then link the other ones to it via "See also links".

Idea: Take the references from the joss paper, they seem great !
