### 2023-10-31 #5987 @ahudde
Dear @lrnv,
Thank you very much for taking the time to help me improve the package `greeks`. I will start working on the issues and come back to you, if I have questions.
