### 2023-10-31 5450d71
Fixed issue  [vignette] Picture too big #8 https://github.com/ahudde/greeks/issues/8
