### 2023-10-31 b12788d
Deleted unused and fully commented function
