+++ src/BS_European_Greek_cpp.cpp
@@ -1,327 +1,0 @@
-#include <Rcpp.h>
-#include <string>
-
-using std::string;
-
-using namespace Rcpp;
-
-//using R::dnorm;
-// using R::pnorm;
-
-// NumericVector BS_European_Greeks(
-//     double initial_price = 100,
-//     double exercise_price = 100,
-//     double r = 0,
-//     double time_to_maturity = 1,
-//     double volatility = 0.3,
-//     double dividend_yield = 0,
-//     std::string payoff = "call",
-//     CharacterVector greeks = {"fair_value", "delta", "vega", "theta", "rho", "epsilon", "lambda", "gamma", "vanna", "charm", "vomma", "veta", "speed"}) {
-//
-//     NumericVector result =
-//       NumericVector::create(_["fair_value"], _["delta"], _["vega"], _["theta"], _["rho"], _["epsilon"], _["lambda"], _["gamma"], _["vanna"], _["charm"], _["vomma"], _["veta"], _["speed"]));
-
-
-//
-//     names(result) <- greek
-//
-//     d1 <- (log(initial_price/exercise_price) +
-//       (r - dividend_yield + (volatility^2)/2) * time_to_maturity) /
-//         (volatility * sqrt(time_to_maturity))
-//
-//     d2 <- d1 - volatility * sqrt(time_to_maturity)
-//
-//     if (payoff == "call") {
-//
-// // option-value
-//
-//       if ("fair_value" %in% greek) {
-//         result["fair_value"] <-
-//           initial_price * exp(-dividend_yield*time_to_maturity) * pnorm(d1) -
-//           exp(-r*time_to_maturity) * exercise_price * pnorm(d2)
-//       }
-//
-// // first-order Greeks
-//
-//       if ("delta" %in% greek) {
-//         result["delta"] <- exp(-dividend_yield*time_to_maturity) * pnorm(d1)
-//       }
-//       if ("vega" %in% greek) {
-//         result["vega"] <-
-//           initial_price*exp(-dividend_yield*time_to_maturity) * dnorm(d1) *
-//           sqrt(time_to_maturity)
-//       }
-//       if ("theta" %in% greek) {
-//         result["theta"] <-
-//           -initial_price*dnorm(d1)*volatility*
-//           exp(-dividend_yield*time_to_maturity)/(2*sqrt(time_to_maturity)) +
-//           dividend_yield*initial_price*pnorm(d1)*
-//           exp(-dividend_yield*time_to_maturity) -
-//           r*exercise_price*exp(-r*time_to_maturity)*pnorm(d2)
-//       }
-//       if ("rho" %in% greek) {
-//         result["rho"] <-
-//           exercise_price*time_to_maturity*exp(-r*time_to_maturity)*pnorm(d2)
-//       }
-//       if ("epsilon" %in% greek) {
-//         result["epsilon"] <-
-//           -initial_price*time_to_maturity*exp(-dividend_yield*time_to_maturity)*
-//           pnorm(d1)
-//       }
-//       if ("lambda" %in% greek) {
-//         result["lambda"] <-
-//           initial_price * (exp(-dividend_yield*time_to_maturity) * pnorm(d1)) /
-//             (initial_price * exp(-dividend_yield*time_to_maturity) * pnorm(d1) -
-//               exp(-r*time_to_maturity) * exercise_price * pnorm(d2))
-//       }
-//
-// // second-order Greeks
-//
-//       if ("gamma" %in% greek) {
-//         result["gamma"] <-
-//           dnorm(d1)*exp(-dividend_yield*time_to_maturity)/
-//             (initial_price*volatility*sqrt(time_to_maturity))
-//       }
[…]
