### 2023-10-31 490d2a2
Issue: https://github.com/ahudde/greeks/issues/6https://github.com/ahudde/greeks/issues/6, Added two examples in the documentation of R/Greeks.R
