### 2023-10-31 48b361e
Issue https://github.com/ahudde/greeks/issues/9: Clarified, that BS_Geometric_Asian_Greeks computes the Greeks in the Black Scholes model.
