+++ R/BS_Geometric_Asian_Greeks.R
@@ -1,7 +1,6 @@
 #' Computes the Greeks of an Geometric Asian Option with classical Call- and
-#' Put-Payoff
+#' Put-Payoff in the Black Scholes model
 #'
-#' TODO: epsilon
 #'
 #' @export
 #'
