### 2023-10-31 9ab9875
Issue https://github.com/ahudde/greeks/issues/9:
Clarified, that BS_Implied_Volatility computes the Greeks in the Black Scholes model.
