+++ man/BS_Implied_Volatility.Rd
@@ -2,7 +2,8 @@
 % Please edit documentation in R/BS_Implied_Volatility.R
 \name{BS_Implied_Volatility}
 \alias{BS_Implied_Volatility}
-\title{Computes the implied volatility for European options via Halley's method.}
+\title{Computes the implied volatility for European options in the Black Scholes
+model via Halley's method.}
 \usage{
 BS_Implied_Volatility(
   option_price,
@@ -58,7 +59,8 @@ Named vector containing the values of the Greeks specified in the
 parameter \code{greek}.
 }
 \description{
-Computes the implied volatility for European options via Halley's method.
+Computes the implied volatility for European options in the Black Scholes
+model via Halley's method.
 }
 \examples{
 BS_Implied_Volatility(option_price = 27, initial_price = 100,
