### 2023-10-31 2639cda
Issue https://github.com/ahudde/greeks/issues/9:
Clarified, that Malliavin_Aisan_Greeks computes Greeks in the Black Scholes model, or in a Jump Diffusion model.
