### 2023-10-31 f7bac3f
Issue https://github.com/ahudde/greeks/issues/9
Clarified, that BS_Geometric_Asian_Greeks computes the Greeks in the Black Scholes model.
Also fixed a bug, so computing Greeks in the Jump Diffusion Model is actually possible.
