+++ R/Greeks.R
@@ -1,4 +1,5 @@
-#' Computes the Greeks of various options
+#' Computes the Greeks of various options in the Black Scholes model or in a
+#' Jump Diffusion model in the cases of Asian Options
 #'
 #' @export
 #'
@@ -8,7 +9,7 @@
 #' @param time_to_maturity - time to maturity in years
 #' @param dividend_yield - dividend yield
 #' @param volatility - volatility of the underlying asset
-#' @param model - the model to be chosen
+#' @param model - the model to be chosen in ("black_scholes", "jump_diffusion")
 #' @param option_type in c("European", "American", "Asian", "Geometric Asian",
 #' "Digital", "Binomial) - the type of option to be considered
 #' @param payoff - in c("call", "put", "cash_or_nothing_call",
@@ -90,16 +91,16 @@ Greeks <-
     else if (tolower(option_type) == "asian" && tolower(model) == "black_scholes" &&
              ("fair_value" %in% greek || "delta" %in% greek || "rho" %in% greek)) {
       return(BS_Malliavin_Asian_Greeks(payoff = payoff,
-                                    greek = greek,
-                                    initial_price = initial_price,
-                                    exercise_price = exercise_price,
-                                    r = r,
-                                    time_to_maturity = time_to_maturity,
-                                    volatility = volatility,
-                                    dividend_yield = dividend_yield))
+                                       greek = greek,
+                                       initial_price = initial_price,
+                                       exercise_price = exercise_price,
+                                       r = r,
+                                       time_to_maturity = time_to_maturity,
+                                       volatility = volatility,
+                                       dividend_yield = dividend_yield))
     }
 
-    else if (tolower(option_type) == "asian" && tolower(model) == "black_scholes") {
+    else if (tolower(option_type) == "asian" && tolower(model) == "jump_diffusion") {
       return(Malliavin_Asian_Greeks(payoff = payoff,
                                     greek = greek,
                                     initial_price = initial_price,
@@ -114,4 +115,4 @@ Greeks <-
       stop("Wrong input or not yet implemented.")
     }
 
-}
+  }
