+++ man/Greeks.Rd
@@ -2,7 +2,8 @@
 % Please edit documentation in R/Greeks.R
 \name{Greeks}
 \alias{Greeks}
-\title{Computes the Greeks of various options}
+\title{Computes the Greeks of various options in the Black Scholes model or in a
+Jump Diffusion model}
 \usage{
 Greeks(
   initial_price,
@@ -44,7 +45,7 @@ Greeks(
 }}
 
 \item{model}{\itemize{
-\item the model to be chosen
+\item the model to be chosen in ("black_scholes", "jump_diffusion")
 }}
 
 \item{option_type}{in c("European", "American", "Asian", "Geometric Asian",
@@ -70,7 +71,8 @@ Named vector containing the values of the Greeks specified in the
 parameter \code{greek}.
 }
 \description{
-Computes the Greeks of various options
+Computes the Greeks of various options in the Black Scholes model or in a
+Jump Diffusion model
 }
 \examples{
 Greeks(initial_price = 100, exercise_price = 120, r = 0.01,
