### 2023-10-31 6b4ff85
Issue #9: Now, the Binomial options pricing model is mentioned, too.
