+++ R/BS_Geometric_Asian_Greeks.R
@@ -1,4 +1,4 @@
-#' Computes the Greeks of an Geometric Asian Option with classical Call- and
+#' Computes the Greeks of a Geometric Asian Option with classical Call- and
 #' Put-Payoff in the Black Scholes model
 #'
 #' @export
