+++ man/BS_Geometric_Asian_Greeks.Rd
@@ -2,7 +2,7 @@
 % Please edit documentation in R/BS_Geometric_Asian_Greeks.R
 \name{BS_Geometric_Asian_Greeks}
 \alias{BS_Geometric_Asian_Greeks}
-\title{Computes the Greeks of an Geometric Asian Option with classical Call- and
+\title{Computes the Greeks of a Geometric Asian Option with classical Call- and
 Put-Payoff in the Black Scholes model}
 \usage{
 BS_Geometric_Asian_Greeks(
@@ -56,7 +56,7 @@ Named vector containing the values of the Greeks specified in the
 parameter \code{greek}.
 }
 \description{
-Computes the Greeks of an Geometric Asian Option with classical Call- and
+Computes the Greeks of a Geometric Asian Option with classical Call- and
 Put-Payoff in the Black Scholes model
 }
 \examples{
