+++ man/BS_European_Greeks.Rd
@@ -2,7 +2,7 @@
 % Please edit documentation in R/BS_European_Greeks.R
 \name{BS_European_Greeks}
 \alias{BS_European_Greeks}
-\title{Computes the Greeks of an European call- or put-option, or of digital options
+\title{Computes the Greeks of a European call- or put-option, or of digital options
 in the Black Scholes model}
 \usage{
 BS_European_Greeks(
@@ -58,7 +58,7 @@ Named vector containing the values of the Greeks specified in the
 parameter \code{greek}.
 }
 \description{
-Computes the Greeks of an European call- or put-option, or of digital options
+Computes the Greeks of a European call- or put-option, or of digital options
 in the Black Scholes model
 }
 \examples{
