+++ R/BS_Malliavin_Asian_Greeks.R
@@ -1,5 +1,5 @@
-#' Computes the Greeks of an Asian option with the Malliavin Monte Carlo
-#' Method in the Black Scholes model
+#' Computes the Greeks of an Asian option with the Malliavin Monte Carlo Method
+#' in the Black Scholes model
 #'
 #' @export
 #'
@@ -16,7 +16,8 @@
 #' @param dividend_yield - dividend yield
 #' @param payoff - the payoff function, either a string in ("call", "put"), or a
 #' function
-#' @param greek - the Greek to be calculated
+#' @param greek - Greeks to be calculated in c("fair_value", "delta", "gamma",
+#' "gamma_kombi", "rho", "theta", "vega")
 #' @param steps - the number of integration steps
 #' @param paths - the number of simulated paths
 #' @param seed - the seed of the random number generator
@@ -124,8 +125,7 @@ BS_Malliavin_Asian_Greeks <- function(
 
   if (length(intersect(
     greek,
-    c("delta", "theta", "vega", "gamma", "gamma_kombi",
-      "rho_d")))) {
+    c("delta", "theta", "vega", "gamma", "gamma_kombi", "rho_d")))) {
     I_1 <- calc_I_1(X, steps, dt)
     I_2 <- calc_I_2(X, steps, dt)
   }
