### 2023-11-01 1d4994c
Bugfix, dependency of dividend_yield
