+++ R/Malliavin_European_Greeks.R
@@ -14,7 +14,8 @@
 #' @param dividend_yield - dividend yield
 #' @param payoff - the payoff function, either a string in ("call", "put",
 #' "digital_call", "digital_put"), or a function
-#' @param greek - the greek to be calculated
+#' @param greek - the Greeks to be calculated in ("fair_value", "delta",
+#' "vega", "theta", "rho", "gamma")
 #' @param model - the model to be chosen
 #' @param paths - the number of simulated paths
 #' @param seed - the seed of the random number generator
@@ -24,8 +25,9 @@
 #' @return Named vector containing the values of the Greeks specified in the
 #' parameter \code{greek}
 #'
-#' @examples Malliavin_European_Greeks(initial_price = 110, exercise_price = 100,
-#' r = 0.02, time_to_maturity = 4.5, dividend_yield = 0.015, volatility = 0.22,
+#' @examples Malliavin_European_Greeks(initial_price = 110,
+#' exercise_price = 100, r = 0.02, time_to_maturity = 4.5,
+#' dividend_yield = 0.015, volatility = 0.22,
 #' greek = c("fair_value", "delta", "rho"), payoff = "put")
 #'
 
@@ -84,14 +86,16 @@ Malliavin_European_Greeks <-
 
   if (model == "Black Scholes") {
     X_T <- initial_price *
-      exp((r-(volatility^2)/2)*time_to_maturity + (volatility*W_T))
+      exp(((r-dividend_yield) - (volatility^2)/2)*time_to_maturity +
+            (volatility*W_T))
     } else {
     print("Unknown model")
     return()
   }
 
   E <- function(weight) {
-    return(exp(-r*time_to_maturity) * mean(payoff(X_T) * weight))
+    return(exp(-(r-dividend_yield)*time_to_maturity) *
+             mean(payoff(X_T) * weight))
   }
 
   if ("fair_value" %in% greek) {
@@ -120,8 +124,8 @@ Malliavin_European_Greeks <-
   if ("theta" %in% greek) {
     result["theta"] <-
       -(W_T^2/(2*time_to_maturity^2) +
-         (r - volatility^2/2)*W_T/(volatility*time_to_maturity) -
-         (1/(2*time_to_maturity) + r)) %>%
+         ((r-dividend_yield) - volatility^2/2)*W_T/(volatility*time_to_maturity) -
+         (1/(2*time_to_maturity) + (r-dividend_yield))) %>%
       E()
   }
 
