+++ R/Malliavin_European_Greeks.R
@@ -11,7 +11,6 @@
 #' @param r - risk-free interest rate
 #' @param time_to_maturity - time to maturity in years
 #' @param volatility - volatility of the underlying asset
-#' @param dividend_yield - dividend yield
 #' @param payoff - the payoff function, either a string in ("call", "put",
 #' "digital_call", "digital_put"), or a function
 #' @param greek - the Greeks to be calculated in ("fair_value", "delta",
@@ -27,8 +26,7 @@
 #'
 #' @examples Malliavin_European_Greeks(initial_price = 110,
 #' exercise_price = 100, r = 0.02, time_to_maturity = 4.5,
-#' dividend_yield = 0.015, volatility = 0.22,
-#' greek = c("fair_value", "delta", "rho"), payoff = "put")
+#' volatility = 0.22, greek = c("fair_value", "delta", "rho"), payoff = "put")
 #'
 
 Malliavin_European_Greeks <-
@@ -37,7 +35,6 @@ Malliavin_European_Greeks <-
            r = 0,
            time_to_maturity = 1,
            volatility = 0.3,
-           dividend_yield = 0,
            payoff = "call",
            greek = c("fair_value", "delta", "vega", "theta", "rho", "gamma"),
            model = "Black Scholes",
@@ -86,7 +83,7 @@ Malliavin_European_Greeks <-
 
   if (model == "Black Scholes") {
     X_T <- initial_price *
-      exp(((r-dividend_yield) - (volatility^2)/2)*time_to_maturity +
+      exp((r - (volatility^2)/2)*time_to_maturity +
             (volatility*W_T))
     } else {
     print("Unknown model")
@@ -94,7 +91,7 @@ Malliavin_European_Greeks <-
   }
 
   E <- function(weight) {
-    return(exp(-(r-dividend_yield)*time_to_maturity) *
+    return(exp(-r*time_to_maturity) *
              mean(payoff(X_T) * weight))
   }
 
@@ -124,8 +121,8 @@ Malliavin_European_Greeks <-
   if ("theta" %in% greek) {
     result["theta"] <-
       -(W_T^2/(2*time_to_maturity^2) +
-         ((r-dividend_yield) - volatility^2/2)*W_T/(volatility*time_to_maturity) -
-         (1/(2*time_to_maturity) + (r-dividend_yield))) %>%
+         (r - volatility^2/2)*W_T/(volatility*time_to_maturity) -
+         (1/(2*time_to_maturity) + r)) %>%
       E()
   }
 
