### 2023-11-01 362cef2
Test Malliavin_European_Greeks, one step towards issue #17
