### 2023-11-01 fccea0f
Issue #17 - improved test coverage
