### 2023-11-02 671eef8
Issue #17, increased test coverage
