### 2023-11-02 74c5fa6
#17, improve test coverage
