### 2023-11-03 119e3ab
Issue #17, increase test coverage
