### 2023-11-03 c5c69d8
#17 increase test coverage
