+++ tests/testthat/test-Malliavin_Asian_Greeks.R
@@ -1,14 +1,9 @@
 test_that("Malliavin_Asian_Greeks is correct", {
 
-  # We check the Greeks by also computing the derivative with finite difference
-  # and comparing the results
+  ## We compare delta - delta_d, vega - vega_d, rho - rho_d, theta - theta_d,
+  ## gamma - gamma_kombi
 
-  number_of_runs <- 2
-
-  definition_of_greeks <-
-    data.frame(greek = "delta", start = "fair_value", param = "initial_price") %>%
-    add_row(greek = "rho", start = "fair_value", param = "r") %>%
-    add_row(greek = "vega", start = "fair_value", param = "volatility")
+  number_of_runs <- 10
 
   error <- numeric(number_of_runs)
 
@@ -20,13 +15,154 @@ test_that("Malliavin_Asian_Greeks is correct", {
     initial_price <- runif(1, 90, 110)
     exercise_price <- runif(1, 90, 110)
     r <- runif(1, -0.01, 0.1)
-    time_to_maturity <- runif(1, 0.2, 6)
+    time_to_maturity <- runif(1, 0.2, 2)
     dividend_yield <- runif(1, 0, 0.1)
     volatility <- runif(1, 0.01, 1)
     model <- "Black_Scholes"
-    payoff <- rep(c("call", "put"), 3)[i]
-    greek <- rep(c("delta"), 2)[i]
-    antithetic <- c(TRUE, FALSE)[i]
+    payoff <- rep(c("call", "put"), number_of_runs)[i]
+    greek <- c("delta", "delta", "vega", "vega", "rho", "rho", "theta", "theta",
+               "gamma", "gamma")[i]
+    greek_d <- c("delta_d", "delta_d", "vega_d", "vega_d", "rho_d", "rho_d",
+               "theta_d", "theta_d", "gamma_kombi", "gamma_kombi")[i]
+    param <- "initial_price"
+    start <- "fair_value"
+
+    Value <-
+      Malliavin_Asian_Greeks(
+        initial_price = initial_price,
+        exercise_price = exercise_price,
+        r = r,
+        time_to_maturity = time_to_maturity,
+        volatility = volatility,
+        dividend_yield = dividend_yield,
+        payoff = payoff,
+        greek = greek,
+        steps = 10,
+        paths = 100000
+      )
+
+    Value_d <-
+      Malliavin_Asian_Greeks(
+        initial_price = initial_price,
+        exercise_price = exercise_price,
+        r = r,
+        time_to_maturity = time_to_maturity,
+        volatility = volatility,
+        dividend_yield = dividend_yield,
+        payoff = payoff,
+        greek = greek_d,
+        steps = 10,
+        paths = 100000
+      )
+
+    error[i] <-
+      min(abs(Value - Value_d)/(abs(Value) + 0.0001),
+          abs(Value - Value_d))
+
+  }
+
+  expect(max(error) < 0.1, "Malliavin_Asian_Greeks: Malliavin - Malliavin_d
+         difference too high")
+
+
+
+  # We check the Greeks by also computing the derivative with finite difference
+  # and comparing the results
+
+  definition_of_greeks <-
+    data.frame(greek = "delta", start = "fair_value", param = "initial_price") %>%
+    add_row(greek = "rho", start = "fair_value", param = "r") %>%
+    add_row(greek = "vega", start = "fair_value", param = "volatility")
+
+  # We start with checking delta
+
+  number_of_runs <- 10
+
+  error <- numeric(number_of_runs)
+
+  set.seed(42)
+
+  for(i in 1:number_of_runs) {
+
[…]
