### 2023-11-03 af61504
#17 increase test coverage
