### 2023-11-03 e845f69
#17 increase test coverage
