### 2023-11-11 f6771ee
#17: Increase test coverage
