### 2023-11-13 anselmhudde/greeks#18 @bahung
[paper] general structure and comments
Hi @ahudde

Below are my comments for your paper current version:
1. Structure: Please add one section to describe the components of greeks, each with visualizations accordingly, especially when one of your paper main contributions is "Also, it is the only R package with included interactive visualization"
2. What is "Shiny app"? If it is a way to display the interactive visualizations, I think it is better to explain the mechanism to produce such visualizations and how easy to interact with them.
3. There are strong claims on the last paragraph in your statement of need section:
- "greeks is used in graduate courses in financial mathematics to provide a better understanding of option prices and Greeks by interactive visualizations."
- "greeks is also suited for financial risk management purposes".
Please cite the appropriate source here or provide evidence to such claims.
4. More related works are needed, especially on Asian option greeks calculations
- Yu, C. and Wang, X., 2023. Quasi-Monte Carlo-based conditional Malliavin method for continuous-time Asian option greeks. Computational Economics, 62(1), pp.325-360.
- Willems, S., 2019. Asian option pricing with orthogonal polynomials. Quantitative Finance, 19(4), pp.605-618.
- Lyuu, Y.D., Teng, H.W., Tseng, Y.T. and Wang, S.X., 2019. A systematic and efficient simulation scheme for the Greeks of financial derivatives. Quantitative Finance, 19(7), pp.1199-1219.
 and how to visualize and interact with computational tools:
- Barchart: https://www.barchart.com/options/options-calculator
- coggit: http://www.coggit.com/tools/american_option_pricing.html
- Options Calculator: https://www.optionseducation.org/toolsoptionquotes/optionscalculator

Bests
