### 2023-11-13 #5987 @bahung
Hi @ahudde, posted a new issue regarding your paper https://github.com/ahudde/greeks/issues/18#issue-1990424869
Please address them so that I can clear Software Paper checklist above, thanks!
