### 2023-11-18 anselmhudde/greeks#18 @ahudde
Dear @bahung,

Thank you very much for your comments!

1. and 2.: Shiny is a framework to program interactive apps, see https://shiny.posit.co/ for more information. The packages `greeks` comes with one shiny app. After you have installed the package, you can run it with `greeks::Greeks_UI()` (see also the description on readme.md). The app should be quite intuitively to operate, please try it yourself.
