### 2023-11-18 anselmhudde/greeks#18 @ahudde
3.1:. I am a professor of Data Science and use this package in my graduate course in financial mathematics.
