### 2023-11-18 anselmhudde/greeks#18 @ahudde
3.2. and 4.: I think that this belongs to Issue #11: I need more references which describe the methods that are implemented in `greeks` Also I should mention a source that describes, how the computation of Greeks is applied in risk management. I will update #11 with this point and work on this.
