### 2023-12-01 anselmhudde/greeks#18 @bahung
Hi @ahudde,

Thank you for your responses,
1 and 2 are for paper so that reading your paper will provide information for readers to at least imagine what they can achieve and by what steps with your app. So, please address it.

It is also related to issue #2

3.1 it is great that the app is used wildly, kindly include a link to the course you are teaching, that I think will be a win for both sides.

That's all for the paper. I will comment for the app in other issues.
