### 2023-12-02 anselmhudde/greeks#19 @ahudde
Improve description of Shiny app
Two comments from @bahung:
- Structure: Please add one section to describe the components of greeks, each with visualizations accordingly, especially when one of your paper main contributions is  "Also, it is the only R package with included interactive visualization"
- What is "Shiny app"? If it is a way to display the interactive visualizations, I think it is better to explain the mechanism to produce such visualizations and how easy to interact with them.
