### 2023-12-02 anselmhudde/greeks#18 @ahudde
Dear @bahung,

For 1. and 2., I opened a new issue #19. I think it's better to address your points with single issues. But I will need more help from you with this issue. I will address this later.

I am currently teaching a masters course in financial mathematics of master students of applied mathematics. At our university, the course sites are restricted to students of the course. The course curriculum includes the Black Scholes Model and the computation of Greeks. We use the shiny app to get a better intuitition of options prices and greeks. Do you want to know the curriculum of the course, or what information is lacking?
