### 2023-12-08 #5987 @csoneson
Ping @bahung

> @bahung, it looks like the issue you mention above has been closed, are you able to continue your review?
