### 2023-12-11 anselmhudde/greeks#18 @bahung
Hi @ahudde
I do not want it personally, what I mentioned in 3. above means that the readers of your paper or the users of your app will objectively see the benefits of using your app. So, please make the supporting evidence seen by readers/users. I also taught these courses back in 2014, and one way to provide evidence is a link to the course curriculum, such as https://platform.onlinelearning.upenn.edu/offering/finance-quantitative-modeling-for-analysts-a0Q2E00000JnjO9UAJ, and you link's content should mention Greeks.
