+++ man/Malliavin_European_Greeks.Rd
@@ -42,7 +42,7 @@ Malliavin_European_Greeks(
 
 \item{payoff}{\itemize{
 \item the payoff function, either a string in ("call", "put",
-"digital_call", "digital_put"), or a function
+"cash_or_nothing_call", "cash_or_nothing_call"), or a function
 }}
 
 \item{greek}{\itemize{
