+++ R/BS_European_Greeks.R
@@ -287,21 +287,21 @@ BS_European_Greeks <-
           expression(
             initial_price *
               exp(-dividend_yield * time_to_maturity) *
-              pnorm(pnorm((log(initial_price / exercise_price) +
-                             (r - dividend_yield + (volatility^2) / 2) *
-                             time_to_maturity) /
-                            (volatility * sqrt(time_to_maturity)))))
+              pnorm(
+                (log(initial_price / exercise_price) +
+                   (r - dividend_yield + (volatility^2) / 2) * time_to_maturity) /
+                  (volatility * sqrt(time_to_maturity))))
       }
 
       if (payoff == "asset_or_nothing_put") {
         fair_value <-
           expression(
             initial_price *
-              exp(-r * time_to_maturity) *
-              pnorm(-pnorm((log(initial_price / exercise_price) +
-                              (r - dividend_yield + (volatility^2) / 2) *
-                              time_to_maturity) /
-                             (volatility * sqrt(time_to_maturity)))))
+              exp(-dividend_yield * time_to_maturity) *
+              pnorm(-(
+                (log(initial_price / exercise_price) +
+                   (r - dividend_yield + (volatility^2) / 2) * time_to_maturity) /
+                  (volatility * sqrt(time_to_maturity)))))
       }
 
       ## option-value
