### 2024-01-31 a12caaf
Added asset-or-nothing digital options
