### 2024-01-31 7b9e5fa
#17: increase test coverage
