+++ tests/testthat/test-Malliavin_European_Greeks.R
@@ -20,7 +20,8 @@ test_that("Malliavin_European_Greeks is correct", {
     volatility <- runif(1, 0.01, 1)
     model <- "Black_Scholes"
     payoff <- sample(
-      c("call", "put", "cash_or_nothing_call", "cash_or_nothing_put"), 1)
+      c("call", "put", "cash_or_nothing_call", "cash_or_nothing_put",
+        "asset_or_nothing_put"), 1)
     antithetic <- sample(c(TRUE, FALSE), 1)
     greek <- sample(Greeks, 1)
 
