### 2024-01-31 43953b9
Refactoring of Malliavin_Geometric_Asian_Greeks.R: Removed unnessary conditions
