+++ R/Malliavin_Geometric_Asian_Greeks.R
@@ -9,7 +9,7 @@
 #'
 #' @param initial_price - initial price of the underlying asset, can also be a
 #' vector
-#' @param exercise_price - strike price of the option
+#' @param exercise_price - strike price of the option, can also be a vector
 #' @param r - risk-free interest rate
 #' @param time_to_maturity - time to maturity in years
 #' @param volatility - volatility of the underlying asset
