### 2024-01-31 40afbcf
#17: increase test coverage
