### 2024-01-31 f118066
refactoring: removing unnessary interval functions
