### 2024-02-01 9c3de72
Dear @bahung,
Thank you very much for the idea to link the course description! The course description mentions Risk-neutral valuation, the Black Scholes Model, and hedging. This of course includes option valuation and Delta-hedging, which use the Greek Delta. But we also covered other Greeks in the course.
