### 2024-02-01 anselmhudde/greeks#18 @ahudde
Dear @bahung,

Thank you very much for the idea to link the course description! I added the link (see [9c3de72](https://github.com/ahudde/greeks/commit/9c3de7223137d231ab44297cd4b0ad5ba43a3b02)). The course description mentions Risk-neutral valuation, the Black Scholes Model, and hedging. This of course includes option valuation and Delta-hedging, which use the Greek Delta. But we also covered other Greeks in the course.
