+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: greeks
 Title: Sensitivities of Prices of Financial Options and Implied Volatilities
-Version: 1.3.2
+Version: 1.3.3
 Authors@R: 
     person(given = "Anselm",
            family = "Hudde",
