+++ R/Greeks.R
@@ -2,6 +2,8 @@
 #' the Black Scholes model or a Jump Diffusion model in the case of Asian
 #' Options, or in the Binomial options pricing model
 #'
+#' @export
+#'
 #' @seealso [BS_European_Greeks()] for option_type = "European
 #' @seealso [Binomial_American_Greeks()] for option_type = "American"
 #' @seealso [BS_Geometric_Asian_Greeks()] for option_type = = "Geometric Asian"
@@ -10,8 +12,6 @@
 #' and model = "black_scholes" and greek in c("fair_value", "delta", "rho",
 #' "vega")
 #' @seealso [Malliavin_Asian_Greeks()] for more general cases of Asian Greeks
-#'
-#' @export
 #'
 #' @param initial_price - initial price of the underlying asset
 #' @param exercise_price - strike price of the option
