+++ man/Implied_Volatility.Rd
@@ -72,3 +72,7 @@ Computes the implied volatility for various options via Newton's method
 Implied_Volatility(15, r = 0.05, option_type = "Asian",
 payoff = "call")
 }
+\seealso{
+\code{\link[=BS_Implied_Volatility]{BS_Implied_Volatility()}} for option_type = "European" and payoff
+in c("call", "put")
+}
