+++ R/BS_Geometric_Asian_Greeks.R
@@ -3,6 +3,11 @@
 #'
 #' @export
 #'
+#' @seealso [Malliavin_Geometric_Asian_Greeks] for the Monte Carlo
+#' implementation which provides digital and custom payoff functions and also
+#' works for the jump diffusion model
+#' @seealso [Greeks_UI] for an interactive visualization
+#'
 #' @import "stats"
 #' @import "Rcpp"
 #' @importFrom "dqrng" "dqrnorm" "dqset.seed"
