+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: greeks
 Title: Sensitivities of Prices of Financial Options and Implied Volatilities
-Version: 1.3.3
+Version: 1.3.4
 Authors@R: 
     person(given = "Anselm",
            family = "Hudde",
