+++ tests/testthat/test-Binomial_American_Greeks.R
@@ -112,16 +112,9 @@ test_that("Binomial_American_Greeks fair_value is correct", {
              volatility = 0.4,
              dividend_yield = 0,
              payoff = "put",
-             greek = c("fair_value", "delta", "vega", "theta", "rho", "epsilon",
-                       "gamma"),
-             steps = 5,
-             eps = 1/10000,
-             control_variate = TRUE) {
+             steps = 5) {
 
       ## the payoff function ##
-
-      # TODO: das ist hässlich
-      payoff_name <- payoff
 
       if (payoff == "call") {
         payoff <- function(x, exercise_price) {
@@ -132,10 +125,6 @@ test_that("Binomial_American_Greeks fair_value is correct", {
           return(max(0, exercise_price - x))
         }
       }
-
-      result <- numeric(length(greek)) * NA
-
-      names(result) <- greek
 
       underlying <- matrix(NA, nrow = steps + 1, ncol = steps + 1)
       american_option_value <- matrix(NA, nrow = steps + 1, ncol = steps + 1)
@@ -157,41 +146,26 @@ test_that("Binomial_American_Greeks fair_value is correct", {
           underlying[i, j] <- down * underlying[i-1, j-1]
         }
       }
-
       # initializing with the prices
       for(i in 1:(steps+1)) {
-        american_option_value[i, steps + 1] <- payoff(underlying[i, steps + 1], exercise_price)
-        european_option_value[i, steps + 1] <- payoff(underlying[i, steps + 1], exercise_price)
+        american_option_value[i, steps + 1] <- exp(-r*steps*dt) * payoff(underlying[i, steps + 1], exercise_price)
+        european_option_value[i, steps + 1] <- exp(-r*steps*dt) * payoff(underlying[i, steps + 1], exercise_price)
       }
 
       for (j in steps:1) {
         for (i in 1:j) {
           american_option_value[i, j] <-
             max(
-              (american_option_value[i, j+1] * p + american_option_value[i+1, j+1] * (1-p)) * exp(-r*dt),
-              payoff(underlying[i, j], exercise_price))
+              (american_option_value[i, j+1] * p + american_option_value[i+1, j+1] * (1-p)),
+              exp(-(r-dividend_yield)*(j-1)*dt) * payoff(underlying[i, j], exercise_price))
           european_option_value[i, j] <-
             (european_option_value[i, j+1] * p + european_option_value[i+1, j+1] * (1-p)) * exp(-r*dt)
         }
       }
 
-      if(control_variate == FALSE) {
-        return(american_option_value[1, 1])
-      } else {
-        return(
-          american_option_value[1, 1] - european_option_value[1, 1] +
-            BS_European_Greeks(initial_price = initial_price,
-                               exercise_price = exercise_price,
-                               r = r,
-                               time_to_maturity = time_to_maturity,
-                               volatility = volatility,
-                               dividend_yield = dividend_yield,
-                               payoff = payoff_name,
-                               greek = "fair_value"))
-      }
+      return(american_option_value[1, 1])
 
     }
-
 
   number_of_runs <- 50
 
@@ -231,14 +205,11 @@ test_that("Binomial_American_Greeks fair_value is correct", {
           volatility = volatility,
           dividend_yield = dividend_yield,
           payoff = payoff,
-          control_variate = FALSE,
           steps = 5
         )
     )
 
   }
-
-  expect(max(error) < 0.1)
+  expect(max(error) < 1e-5)
 
 })
-
