+++ man/Greeks.Rd
@@ -2,8 +2,8 @@
 % Please edit documentation in R/Greeks.R
 \name{Greeks}
 \alias{Greeks}
-\title{Computes the Greeks of various options in the Black Scholes model or both in
-the Black Scholes model or a Jump Diffusion model in the case of Asian
+\title{Computes the Greeks of various options in the Black Scholes model or
+both in the Black Scholes model or a Jump Diffusion model in the case of Asian
 Options, or in the Binomial options pricing model}
 \usage{
 Greeks(
@@ -78,9 +78,65 @@ Named vector containing the values of the Greeks specified in the
 parameter \code{greek}.
 }
 \description{
-Computes the Greeks of various options in the Black Scholes model or both in
-the Black Scholes model or a Jump Diffusion model in the case of Asian
-Options, or in the Binomial options pricing model
+Greeks are derivatives of the option value with respect to
+underlying parameters.
+For instance, the Greek
+\eqn{\Delta = \frac{\partial \text{fair\_value}}{\partial \text{initial\_price}}}
+(Delta) measures how the price of an option changes with a minor change in
+the underlying asset's price, while
+\eqn{\Gamma = \frac{\partial \text{fair\_value}}{\partial \text{initial\_price}}}
+(Gamma) measures how \eqn{\Delta} itself changes as the price of the
+underlying asset shifts.
+Greeks can be computed for different types of options:
+For
+\itemize{
+\item \strong{European Greeks} see also \link{BS_European_Greeks} and \link{Malliavin_European_Greeks}
+\item \strong{American Greeks} see also \link{Binomial_American_Greeks}
+\item \strong{Asian Greeks} see also \link{BS_Malliavin_Asian_Greeks} and \link{Malliavin_Asian_Greeks}
+\item \strong{Geometric Asian Greeks} see also \link{BS_Geometric_Asian_Greeks} and \link{Malliavin_Asian_Greeks}
+}
+
+The Greeks are defined as the following partial derivatives of the option
+value:
+\itemize{
+\item \code{Delta} = \eqn{\Delta = \frac{\partial \text{fair\_value}}{\partial \text{initial\_price}}}, the derivative with respect to the price of the
+underlying asset
+\item \code{Vega} = \eqn{\mathcal{V} = \frac{\partial \text{fair\_value}}{\partial \text{volatility}}},
+the derivative with respect to the volatility
+\item \code{Theta} = \eqn{\Theta = -\frac{\partial \text{fair\_value}}{\partial \text{time\_to\_maturity}}},
+the negative derivative with respect to the time until expiration of the
+option
+\item \code{rho} = \eqn{\rho = \frac{\partial \text{fair\_value}}{\partial r}},
+the derivative with respect to the risk-free interest rate
+\item \code{Epsilon} = \eqn{\epsilon = \frac{\partial \text{fair\_value}}{\partial \text{time\_to\_maturity}}},
+the derivative with respect to the dividend yield of the underlying asset
+\item \code{Lambda} = \eqn{\lambda = \Delta \times \frac{\text{initial\_price}}{\text{exercise\_price}}}
+\item \code{Gamma} = \eqn{\Gamma = \frac{\partial^2 \text{fair\_value}}{\partial \text{initial\_price}^2}}, the second derivative with respect to the price of
+the underlying asset
[…]
