### 2024-02-23 anselmhudde/greeks#5 @ahudde
Dear @lrnv : I worked on the description field of the function `Greeks`. See the file `greeks.pdf` for the resulting reference manual. Is this going in the right direction? Yours, Anselm
