### 2024-02-23 anselmhudde/greeks#5 @lrnv
Yes, the description of the Greeks function is definitely better, great job. There are still functions without much description though, so you have to either complete their descriptions either refer to other parts of the docs that do (like the now-very-descriptive Greeks docs)
