### 2024-02-23 anselmhudde/greeks#5 @ahudde
Great, thanks for the feedback, I will work on the other functions, too.
