+++ man/BS_Malliavin_Asian_Greeks.Rd
@@ -72,8 +72,7 @@ Named vector containing the values of the Greeks specified in the
 parameter \code{greek}.
 }
 \description{
-Computes the Greeks of an Asian option with the Malliavin Monte Carlo Method
-in the Black Scholes model
+For a descriptoin see \link{Malliavin_Asian_Greeks}
 }
 \examples{
 BS_Malliavin_Asian_Greeks(initial_price = 110, exercise_price = 100,
