+++ man/BS_Geometric_Asian_Greeks.Rd
@@ -56,8 +56,12 @@ Named vector containing the values of the Greeks specified in the
 parameter \code{greek}.
 }
 \description{
-Computes the Greeks of a Geometric Asian Option with classical Call- and
-Put-Payoff in the Black Scholes model
+In contrast to a Asian options (see \link{Malliavin_Aisan_Greeks}), geometric
+Asian options evaluate the gemetric average
+\eqn{\exp \left( \frac{1}{T} \int_0^T \ln S_t dt \right)}, where
+\eqn{S_t} is the price of the underlying asset at time \eqn{t} and \eqn{T} is
+the time-to-maturity of the option.
+For more details on the definition of Greeks see the function \link{Greeks}.
 }
 \examples{
 BS_Geometric_Asian_Greeks(initial_price = 110, exercise_price = 100,
