+++ man/Binomial_American_Greeks.Rd
@@ -2,8 +2,8 @@
 % Please edit documentation in R/Binomial_American_Greeks.R
 \name{Binomial_American_Greeks}
 \alias{Binomial_American_Greeks}
-\title{Computes the Greeks of an American call- or put-option with the Binomial
-options pricing model}
+\title{Computes the Greeks of an American call- or put-option with the
+Binomial options pricing model}
 \usage{
 Binomial_American_Greeks(
   initial_price = 100,
@@ -65,8 +65,7 @@ Named vector containing the values of the Greeks specified in the
 parameter \code{greek}.
 }
 \description{
-Computes the Greeks of an American call- or put-option with the Binomial
-options pricing model
+For details see the function \link{Greeks}
 }
 \examples{
 Binomial_American_Greeks(initial_price = 100, exercise_price = 100,
