+++ R/Binomial_American_Greeks.R
@@ -1,6 +1,6 @@
 #' @title
-#' Computes the Greeks of an American call- or put-option with the
-#' Binomial options pricing model
+#' Computes the Greeks of an American call- or put-option with the Binomial
+#' options pricing model
 #'
 #' @description
 #' For details see the function [Greeks]
