+++ R/Malliavin_European_Greeks.R
@@ -1,5 +1,11 @@
+#' @title
 #' Computes the Greeks of a European option with the Malliavin Monte Carlo
 #' Method in the Black Scholes model
+#'
+#' @description
+#' For details on the definition of Greeks see [Greeks].
+#' For a description of Malliavin Monte Carlo Methods for Greeks see for example
+#'(Hudde & Rüschendorf, 2023)
 #'
 #' @export
 #'
@@ -31,6 +37,9 @@
 #' @examples Malliavin_European_Greeks(initial_price = 110,
 #' exercise_price = 100, r = 0.02, time_to_maturity = 4.5,
 #' volatility = 0.22, greek = c("fair_value", "delta", "rho"), payoff = "put")
+#'
+#' @references
+#' Hudde, A., & Rüschendorf, L. (2023). European and Asian Greeks for Exponential Lévy Processes. Methodol Comput Appl Probab, 25 (39). [https://doi.org/10.1007/s11009-023-10014-5](https://doi.org/10.1007/s11009-023-10014-5)
 #'
 
 Malliavin_European_Greeks <-
