+++ man/BS_Geometric_Asian_Greeks.Rd
@@ -56,8 +56,8 @@ Named vector containing the values of the Greeks specified in the
 parameter \code{greek}.
 }
 \description{
-In contrast to a Asian options (see \link{Malliavin_Asian_Greeks}), geometric
-Asian options evaluate the gemetric average
+In contrast to Asian options (see \link{Malliavin_Asian_Greeks}), geometric Asian
+options evaluate the geometric average
 \eqn{\exp \left( \frac{1}{T} \int_0^T \ln S_t dt \right)}, where
 \eqn{S_t} is the price of the underlying asset at time \eqn{t} and \eqn{T} is
 the time-to-maturity of the option.
