+++ man/BS_Implied_Volatility.Rd
@@ -59,8 +59,11 @@ Named vector containing the values of the Greeks specified in the
 parameter \code{greek}.
 }
 \description{
-Computes the implied volatility for European put- and call options in the
-Black Scholes model via Halley's method.
+For the definition of \emph{implied volatility} see the function
+\link{Implied_Volatility}.
+\link{BS_Implied_Volatility} offers a very fast implementation for European put-
+and call options applying Halley's method (see
+\href{https://en.wikipedia.org/wiki/Halley\%27s_method}{en.wikipedia.org/wiki/Halley\%27s_method}).
 }
 \examples{
 BS_Implied_Volatility(option_price = 27, initial_price = 100,
