### 2024-02-25 #5987 @ahudde
Dear @csoneson,
Thank you very much for your inquiry and your patience! Since last time, we managed to resolve four issues, and I’m currently working on [#5](https://github.com/ahudde/greeks/issues/5). While working on the issues (especially on increasing the test coverage) I also found some bugs which I solved, but took some time.
On [#18](https://github.com/ahudde/greeks/issues/18), I’m still waiting for a response for one detail, and I probably need some help with [#19](https://github.com/ahudde/greeks/issues/19), I will ask @bahung.
I feel, that the comments really help to significantly improve the package, I hope to finish all comments soon.
Yours,
Anselm
