+++ R/Binomial_American_Greeks.R
@@ -3,7 +3,11 @@
 #' options pricing model
 #'
 #' @description
-#' For details see the function [Greeks]
+#' In contract to European Options, American options can a executed at any time
+#' until the expiration date.
+#' For more details on the definition of Greeks in general see [Greeks].
+#' This functions computes Greeks of American put- and call options in the
+#' binomial option pricing model (see (Hull, 2022)).
 #'
 #' @export
 #'
@@ -34,6 +38,9 @@
 #' r = 0, time_to_maturity = 1, volatility = 0.3, dividend_yield = 0,
 #' payoff = "call", greek = c("fair_value", "delta", "vega", "theta", "rho",
 #' "epsilon", "gamma"), steps = 20)
+#'
+#'@references
+#' Hull, J. C. (2022). Options, futures, and other derivatives (11th Edition). Pearson
 #'
 
 Binomial_American_Greeks <-
