+++ man/BS_Geometric_Asian_Greeks.Rd
@@ -61,7 +61,7 @@ options evaluate the geometric average
 \eqn{\exp \left( \frac{1}{T} \int_0^T \ln S_t dt \right)}, where
 \eqn{S_t} is the price of the underlying asset at time \eqn{t} and \eqn{T} is
 the time-to-maturity of the option.
-For more details on the definition of Greeks see the function \link{Greeks}.
+For more details on the definition of Greeks see \link{Greeks}.
 }
 \examples{
 BS_Geometric_Asian_Greeks(initial_price = 110, exercise_price = 100,
