+++ R/Greeks_UI.R
@@ -11,7 +11,11 @@
 #' some cases, the set of Greeks is limited.
 #' On the y-Axis, the option value resp. the value of the greeks are displayed,
 #' for the x-axis, several parameters like `initial_price` or `time_to_maturity`
-#' are possible
+#' are possible.
+#'
+#' \if{latex}{
+#' \figure{../vignettes/Greeks_UI.png}{The interdface of the functions}
+#' }
 #'
 #' @export
 #'
