+++ R/Malliavin_Asian_Greeks.R
@@ -12,7 +12,7 @@
 #' For more details on the definition of Greeks in general see [Greeks].
 #'
 #' For a description of Malliavin Monte Carlo Methods for Greeks see for example
-#' (Hudde & Rüschendorf, 2023)
+#' (Hudde & Rüschendorf, 2023).
 #'
 #' @export
 #'
