+++ man/Malliavin_European_Greeks.Rd
@@ -75,7 +75,7 @@ parameter \code{greek}
 \description{
 For details on the definition of Greeks see \link{Greeks}.
 For a description of Malliavin Monte Carlo Methods for Greeks see for example
-(Hudde & Rüschendorf, 2023)
+(Hudde & Rüschendorf, 2023).
 }
 \examples{
 Malliavin_European_Greeks(initial_price = 110,
