+++ man/Malliavin_Geometric_Asian_Greeks.Rd
@@ -98,14 +98,18 @@ Named vector containing the values of the Greeks specified in the
 parameter \code{greek}.
 }
 \description{
-Computes the Greeks of a geometric Asian option with the Malliavin Monte
-Carlo Method in the Black Scholes- or Jump diffusion model
+For the definition of geometric Asian options see
+\link{BS_Geometric_Asian_Greeks}, and for a description of the Malliavin Monte
+Carlo Method for Greeks see for example (Hudde & Rüschendorf, 2023).
 }
 \examples{
 Malliavin_Asian_Greeks(initial_price = 110, exercise_price = 100,
 r = 0.02, time_to_maturity = 4.5, dividend_yield = 0.015, volatility = 0.22,
 greek = c("fair_value", "delta", "rho"), payoff = "put")
 
+}
+\references{
+Hudde, A., & Rüschendorf, L. (2023). European and Asian Greeks for Exponential Lévy Processes. Methodol Comput Appl Probab, 25 (39). \url{https://doi.org/10.1007/s11009-023-10014-5}
 }
 \seealso{
 \link{BS_Geometric_Asian_Greeks} for exact and fast computation in the
