+++ man/Binomial_American_Greeks.Rd
@@ -65,7 +65,7 @@ Named vector containing the values of the Greeks specified in the
 parameter \code{greek}.
 }
 \description{
-In contract to European Options, American options can a executed at any time
+In contract to European Options, American options can be executed at any time
 until the expiration date.
 For more details on the definition of Greeks in general see \link{Greeks}.
 This functions computes Greeks of American put- and call options in the
