+++ R/Malliavin_Geometric_Asian_Greeks.R
@@ -3,9 +3,16 @@
 #' Carlo Method in the Black Scholes- or Jump diffusion model
 #'
 #' @description
-#' For the definition of geometric Asian options see
-#' [BS_Geometric_Asian_Greeks], and for a description of the Malliavin Monte
-#' Carlo Method for Greeks see for example (Hudde & Rüschendorf, 2023).
+#' In contrast to Asian options (see [Malliavin_Asian_Greeks]), geometric Asian
+#' options evaluate the geometric average
+#' \eqn{\exp \left( \frac{1}{T} \int_0^T \ln S_t dt \right)}, where
+#' \eqn{S_t} is the price of the underlying asset at time \eqn{t} and \eqn{T} is
+#' the time-to-maturity of the option (see
+#'
+#' [en.wikipedia.org/wiki/Asian_option#European_Asian_call_and_put_options_with_geometric_averaging](https://en.wikipedia.org/wiki/Asian_option#European_Asian_call_and_put_options_with_geometric_averaging)).
+#' For more details on the definition of Greeks see [Greeks], and for a
+#' description of the Malliavin Monte Carlo Method for Greeks see for example
+#' (Hudde & Rüschendorf, 2023).
 #'
 #' @export
 #'
