+++ R/BS_Geometric_Asian_Greeks.R
@@ -3,12 +3,10 @@
 #' Put-Payoff in the Black Scholes model
 #'
 #' @description
-#' In contrast to Asian options (see [Malliavin_Asian_Greeks]), geometric Asian
-#' options evaluate the geometric average
-#' \eqn{\exp \left( \frac{1}{T} \int_0^T \ln S_t dt \right)}, where
-#' \eqn{S_t} is the price of the underlying asset at time \eqn{t} and \eqn{T} is
-#' the time-to-maturity of the option.
-#' For more details on the definition of Greeks see [Greeks].
+#' For the definition of geometric Asian options see
+#' [Malliavin_Geometric_Asian_Greeks].
+#' [BS_Geometric_Asian_Greeks] offers a fast and exaction computation of
+#' Geometric Asian Greeks.
 #'
 #' @export
 #'
