+++ R/BS_Implied_Volatility.R
@@ -6,6 +6,7 @@
 #' [Implied_Volatility].
 #' [BS_Implied_Volatility] offers a very fast implementation for European put-
 #' and call options applying Halley's method (see
+#'
 #' [en.wikipedia.org/wiki/Halley%27s_method](https://en.wikipedia.org/wiki/Halley%27s_method)).
 #'
 #' @export
