+++ man/Malliavin_Geometric_Asian_Greeks.Rd
@@ -98,9 +98,16 @@ Named vector containing the values of the Greeks specified in the
 parameter \code{greek}.
 }
 \description{
-For the definition of geometric Asian options see
-\link{BS_Geometric_Asian_Greeks}, and for a description of the Malliavin Monte
-Carlo Method for Greeks see for example (Hudde & Rüschendorf, 2023).
+In contrast to Asian options (see \link{Malliavin_Asian_Greeks}), geometric Asian
+options evaluate the geometric average
+\eqn{\exp \left( \frac{1}{T} \int_0^T \ln S_t dt \right)}, where
+\eqn{S_t} is the price of the underlying asset at time \eqn{t} and \eqn{T} is
+the time-to-maturity of the option (see
+
+\href{https://en.wikipedia.org/wiki/Asian_option#European_Asian_call_and_put_options_with_geometric_averaging}{en.wikipedia.org/wiki/Asian_option#European_Asian_call_and_put_options_with_geometric_averaging}).
+For more details on the definition of Greeks see \link{Greeks}, and for a
+description of the Malliavin Monte Carlo Method for Greeks see for example
+(Hudde & Rüschendorf, 2023).
 }
 \examples{
 Malliavin_Asian_Greeks(initial_price = 110, exercise_price = 100,
