### 2024-02-26 anselmhudde/greeks#5 @ahudde
Dear @lrnv, how do you like the documentation now (if you could a glance, it has grown a bit :-))?

Best, Anselm
