### 2024-02-26 anselmhudde/greeks#5 @lrnv
From reading the commits that are referenced here, yeah this is going great.

From reading the documentation itself (the .pdf) it looks a lot better. Great job @ahudde :)

You can close this ticket
