+++ R/Greeks_UI.R
@@ -14,7 +14,7 @@
 #' are possible.
 #'
 #' \if{latex}{
-#' \figure{../vignettes/Greeks_UI.png}{The interdface of the functions}
+#' \figure{Greeks_UI.png}{The interface of the functions}
 #' }
 #'
 #' @export
