### 2024-03-03 anselmhudde/greeks#19 @ahudde
Dear @bahung,
Please have a look at the updated the to include a description of the shiny app.
Furthermore, I updated the reference manual [greeks.pdf](https://github.com/ahudde/greeks/blob/main/greeks.pdf), which you can find on this repository, or on cran: https://cran.r-project.org/web/packages/greeks/greeks.pdf.
What do you think of the description of the function Greeks_UI()? Is this enough?

Yours,
Anselm
