### 2024-03-03 anselmhudde/greeks#18 @ahudde
Dear @bahung,
Does this satisfy point 3 for you?
With best regards,
Anselm
