+++ paper.bib
@@ -7,6 +7,35 @@
   doi = {10.1007/s11009-023-10014-5},
 }
 
+@article{Kou:2002,
+  Author = {S. G. Kou},
+  Title = {A Jump-Diffusion Model for Option Pricing},
+  Journal = {Management Science},
+  Year = {2002},
+  Volume = {48(8)},
+  doi = {10.1287/mnsc.48.8.1086.166},
+}
+
+@article{Lyuu:2019,
+  author = {Yuh-Dauh Lyuu, Huei-Wen Teng, Yao-Te Tseng and Sheng-Xiang Wang},
+  title = {A systematic and efficient simulation scheme for the Greeks of financial derivatives},
+  journal = {Quantitative Finance},
+  volume = {19},
+  number = {7},
+  pages = {1199-1219},
+  year = {2019},
+  doi = {10.1080/14697688.2018.1562196},
+}
+
+@article{Angus:1999,
+  author = {Angus, John E.},
+  title = {A note on pricing Asian derivatives with continuous geometric averaging},
+  journal = {Journal of Futures Markets},
+  volume = {19(7)},
+  pages = {845–858},
+  year = {1999},
+  doi = {10.1002/(SICI)1096-9934(199910)19:7<845::AID-FUT6>3.0.CO;2-D},
+}
 
 @manual{derivmkts,
   Title = {derivmkts: Functions and R Code to Accompany Derivatives Markets},
@@ -40,7 +69,6 @@
   journal = {GitHub repository},
   url = {https://github.com/arfon/fidgit}
 }
-
 
 @MISC{Quantlib,
 author = "Ferdinando Ametrano and Luigi Ballabio",
