### 2024-03-05 anselmhudde/greeks#19 @bahung
Hi @ahudde,

I saw the updates regarding the UI as follows:
"For performance reasons, just the Black-Scholes model is possible, and for some cases, the set of
Greeks is limited. On the y-Axis, the option value resp. the value of the greeks are displayed, for
the x-axis, several parameters like initial_price or time_to_maturity are possible."

It would be nice if you can be more specific:
1.  Why only BS is available?
2.  For "and for some cases, the set of Greeks is limited" --> Does the set depend on the options types? If so, please mention that.

After you finish, I think you can close this issues.

Thank you for your updates,
BaHung
