### 2024-03-05 anselmhudde/greeks#18 @bahung
Hi @ahudde,

With "greeks is used in graduate courses in financial mathematics to provide a better understanding of option prices and Greeks by interactive visualizations." and the page at the link you provided https://www.hs-koblenz.de/fileadmin/media/fb_mathematik-und-technik/Pruefungsamt-Dokumente/Modulhandbuch_Ma-AM_aktuell.pdf#page=32, it seems there is no explicitly mention about your greeks software being used to calculate Greeks. If it is not possible to provide evidence to such claim, please consider rewrite it to make this statement clear.

Bests
