### 2024-03-05 ae3bfdd
#19 Improve description of shiny app
