### 2024-03-05 anselmhudde/greeks#19 @ahudde
@bahung: Thank you. I clarified this sentence.
