### 2024-03-05 anselmhudde/greeks#18 @ahudde
@bahung : I reformulated this sentence. What do you think? Yours, Anselm
