### 2024-03-09 40437fa
Minor changes to paper and bibliography for increased consistency
