+++ paper.bib
@@ -1,7 +1,7 @@
 @article{Hudde:2023,
   Author = {{Hudde}, A. and {Rüschendorf}, L.},
-  Title = {European and Asian Greeks for Exponential Lévy Processes},
-  Journal = {Methodol Comput Appl Probab},
+  Title = {European and {Asian} {Greeks} for Exponential {Lévy} Processes},
+  Journal = {Methodology and Computing in Applied Probability},
   Year = {2023},
   volume = {25 (39)},
   doi = {10.1007/s11009-023-10014-5},
@@ -18,7 +18,7 @@
 
 @article{Lyuu:2019,
   author = {Yuh-Dauh Lyuu, Huei-Wen Teng, Yao-Te Tseng and Sheng-Xiang Wang},
-  title = {A systematic and efficient simulation scheme for the Greeks of financial derivatives},
+  title = {A systematic and efficient simulation scheme for the {Greeks} of financial derivatives},
   journal = {Quantitative Finance},
   volume = {19},
   number = {7},
@@ -29,7 +29,7 @@
 
 @article{Angus:1999,
   author = {Angus, John E.},
-  title = {A note on pricing Asian derivatives with continuous geometric averaging},
+  title = {A note on pricing {Asian} derivatives with continuous geometric averaging},
   journal = {Journal of Futures Markets},
   volume = {19(7)},
   pages = {845–858},
@@ -38,7 +38,7 @@
 }
 
 @manual{derivmkts,
-  Title = {derivmkts: Functions and R Code to Accompany Derivatives Markets},
+  Title = {derivmkts: Functions and {R} Code to Accompany Derivatives Markets},
   Author = {Robert McDonald},
   Year = {2022},
   Note = {R package version 0.2.5},
@@ -54,7 +54,7 @@
 }
 
 @manual{RQuantLib,
-  Title = {RQuantLib: R Interface to the 'QuantLib' Library},
+  Title = {RQuantLib: {R} Interface to the {'QuantLib'} Library},
   Author = {Eddelbuettel, Dirk  and Nguyen, Khanh and Leitch, Terry},
   Year = {2023},
   Note = {R package version 0.4.19},
@@ -63,7 +63,7 @@
 
 @misc{fidgit,
   Author = {A. M. Smith and K. Thaney and M. Hahnel},
-  Title = {Fidgit: An ungodly union of GitHub and Figshare},
+  Title = {Fidgit: An ungodly union of {GitHub} and {Figshare}},
   Year = {2020},
   publisher = {GitHub},
   journal = {GitHub repository},
