+++ paper.bib
@@ -17,7 +17,7 @@
 }
 
 @article{Lyuu:2019,
-  author = {Yuh-Dauh Lyuu, Huei-Wen Teng, Yao-Te Tseng and Sheng-Xiang Wang},
+  author = {Lyuu, Yuh-Dauh and Teng, Huei-Wen and Tseng, Yao-Te and Wang, Sheng-Xiang},
   title = {A systematic and efficient simulation scheme for the {Greeks} of financial derivatives},
   journal = {Quantitative Finance},
   volume = {19},
