### 2024-03-09 #5987 @csoneson
👋🏻 @ahudde - I have gone through your submission and overall it looks good; I made a small [PR](https://github.com/ahudde/greeks/pull/21) with some suggestions for the paper and bibliography. Please check and merge if you agree - you can then generate a new proof with `@editorialbot generate pdf`. I will make another post below with the other remaining steps.
