### 2024-03-09 9ce3861
https://github.com/openjournals/joss-reviews/issues/5987: Version 1.4.0 for publication in JOSS
