+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: greeks
 Title: Sensitivities of Prices of Financial Options and Implied Volatilities
-Version: 1.3.5
+Version: 1.4.0
 Authors@R: 
     person(given = "Anselm",
            family = "Hudde",
