### 2024-03-11 #5987 @csoneson
Great, thanks @ahudde. A couple of small things:

- For the archiving, please first create a release in your GitHub repository, named according to the current version, and then archive this on Zenodo (rather than uploading a tar.gz file). See [here](https://docs.github.com/en/repositories/releasing-projects-on-github/managing-releases-in-a-repository) for instructions.
- Could you remove the `equal-contrib: yes` line from your manuscript file (since there is only one author)? Sorry, I missed this before - for some reason I didn't see this when rendering the manuscript offline.
