### 2024-03-11 95588ca
https://github.com/openjournals/joss-reviews/issues/5987: removed `equal-contrib: yes` from paper.md
