### 2024-03-12 #5987 @csoneson
@ahudde Thanks, the release looks good. Could you please just update the title, author list and license in the Zenodo metadata so that they match those of the paper, then I think we're ready to move forward.
