### 2024-03-12 #5987 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
