### 2024-03-12 6bd4810
paper.md: two grammatical error corrected
Two grammatical errors corrected
