### 2024-03-12 #5987 @csoneson
Sure - just make a new post with only the editorialbot command (it has to be the first command in the post, otherwise it's not recognized)
