### 2024-03-13 #5987 @ahudde
Dear @csoneson,
Thank you very much for you patience, and sorry for the inconveniences. I think the pull-request https://github.com/openjournals/joss-papers/pull/5117 is deleted because I recompiled the pdf. Do I have to publish a new release now, or does it suffice if you recommend-accept again?

Yours, Anselm
