### 2024-03-13 #5987 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
