### 2024-03-16 #5987 @ahudde
Dear @oliviaguest,
I double-checked all DOIs. Options, futures, and other derivatives does indeed not have a DOI.
