### 2024-03-19 #5987 @oliviaguest
Huge thanks to the reviewers, @lrnv, @bahung, and editor, @csoneson! ✨  JOSS appreciates your work and effort. ✨ Also, big congratulations to the authors: @ahudde! 🥳 🍾
